+++ src/rulesDefinition/index.js
@@ -1,0 +1,3 @@
+import rulesDefinition from './rulesDefinition';
+
+export default rulesDefinition;
