+++ src/transform/index.js
@@ -1,0 +1,3 @@
+import transform from './transform';
+
+export default transform;
